+++ src/System.CommandLine/Help/HelpOption.cs
@@ -1,6 +1,9 @@
 ﻿// Copyright (c) .NET Foundation and contributors. All rights reserved.
 // Licensed under the MIT license. See LICENSE file in the project root for full license information.
 
+using System.CommandLine.Invocation;
+using System.CommandLine.IO;
+
 namespace System.CommandLine.Help
 {
     internal class HelpOption : Option<bool>
@@ -33,14 +36,22 @@ namespace System.CommandLine.Help
 
         internal override bool IsGreedy => false;
 
-        public override bool Equals(object? obj)
-        {
-            return obj is HelpOption;
-        }
+        public override bool Equals(object? obj) => obj is HelpOption;
 
-        public override int GetHashCode()
+        public override int GetHashCode() => typeof(HelpOption).GetHashCode();
+
+        internal static void Handler(InvocationContext context)
         {
-            return typeof(HelpOption).GetHashCode();
+            var output = context.Console.Out.CreateTextWriter();
+
+            var helpContext = new HelpContext(context.BindingContext.HelpBuilder,
+                                              context.ParseResult.CommandResult.Command,
+                                              output,
+                                              context.ParseResult);
+
+            context.BindingContext
+                   .HelpBuilder
+                   .Write(helpContext);
         }
     }
 }
