--- src/System.CommandLine/Help/HelpResult.cs
@@ -1,25 +1,0 @@
-﻿// Copyright (c) .NET Foundation and contributors. All rights reserved.
-// Licensed under the MIT license. See LICENSE file in the project root for full license information.
-
-using System.CommandLine.Invocation;
-using System.CommandLine.IO;
-
-namespace System.CommandLine.Help
-{
-    internal static class HelpResult
-    {
-        internal static void Apply(InvocationContext context)
-        {
-            var output = context.Console.Out.CreateTextWriter();
-
-            var helpContext = new HelpContext(context.BindingContext.HelpBuilder,
-                                              context.ParseResult.CommandResult.Command,
-                                              output,
-                                              context.ParseResult);
-
-            context.BindingContext
-                   .HelpBuilder
-                   .Write(helpContext);
-        }
-    }
-}
