+++ src/System.CommandLine/ParseResult.cs
@@ -29,12 +29,14 @@ namespace System.CommandLine
             List<Token> tokens,
             IReadOnlyList<Token>? unmatchedTokens,
             List<ParseError>? errors,
-            string? commandLineText = null)
+            string? commandLineText = null,
+            ICommandHandler? handler = null)
         {
             Parser = parser;
             _rootCommandResult = rootCommandResult;
             CommandResult = commandResult;
             _directives = directives;
+            _handler = handler;
 
             // skip the root command when populating Tokens property
             if (tokens.Count > 1)
